+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 ### Changed
+- Updated `ServiceIdentifier` to rely on `Function` instead of `Abstract<T>`.
 
 ### Fixed